requirements.txt
requests
biopython==1.79
pandas